Renaming component files
